+++ src/components/details-view/details-view.js
@@ -31,6 +31,7 @@ const cx = classNames.bind(styles);
  */
 class DetailsView extends Component {
   static propTypes = {
+    accordionHeaderLoading: PropTypes.bool.isRequired,
     actionMenu: PropTypes.oneOfType([
       PropTypes.func,
       PropTypes.node,
@@ -157,6 +158,7 @@ class DetailsView extends Component {
       onListToggle,
       ariaRole,
       bodyAriaRole,
+      accordionHeaderLoading,
     } = this.props;
 
     const isListAccordionOpen = sections && sections[listSectionId];
@@ -208,9 +210,15 @@ class DetailsView extends Component {
               data-test-eholdings-details-view-list={type}
             >
               <Accordion
-                header={AccordionListHeader}
+                header={props => (
+                  <AccordionListHeader
+                    {...props}
+                    isLoading={accordionHeaderLoading}
+                  />
+                )}
                 headerProps={{
                   resultsLength,
+                  'data-testid': `accordion-toggle-button-${listSectionId}`,
                 }}
                 label={(
                   <Headline